feat: Refactor HSM and HSMV2 components to utilize helper functions for button templates and validation schema
