+++ src/containers/HSM/HSM.helper.ts
@@ -1,3 +1,5 @@
+import * as Yup from 'yup';
+
 import { CALL_TO_ACTION, MEDIA_MESSAGE_TYPES, QUICK_REPLY, WHATSAPP_FORM } from 'common/constants';
 
 export interface CallToActionTemplate {
@@ -122,4 +124,312 @@ export const getExampleValue = (example: string) => {
   }
 
   return variables;
+};
+
+export const regexForShortcode = /^[a-z0-9_]+$/g;
+
+export const buttonTypes: any = {
+  QUICK_REPLY: { value: '' },
+  CALL_TO_ACTION: { type: 'phone_number', title: '', value: '' },
+  WHATSAPP_FORM: { type: 'whatsapp_form', form_id: '', text: '', navigate_screen: '' },
+};
+
+// splits a body string into the message and any trailing "| [button, ...]" gupshup markup.
+export const getTemplateAndButton = (text: string) => {
+  const exp = /(\|\s\[)|(\|\[)/;
+  const areButtonsPresent = text.search(exp);
+
+  let message: any = text;
+  let buttons: any = null;
+
+  if (areButtonsPresent !== -1) {
+    buttons = text.substr(areButtonsPresent);
+    message = text.substr(0, areButtonsPresent);
+  }
+
+  return { message, buttons };
+};
+
+export interface ButtonTemplatePayloadArgs {
+  templateButtons: Array<any>;
+  templateType: any;
+  body: string;
+  variables: Array<any>;
+}
+
+// builds the buttons/example/body payload fragment for whichever button type is selected.
+export const getButtonTemplatePayload = ({
+  templateButtons,
+  templateType,
+  body,
+  variables,
+}: ButtonTemplatePayloadArgs) => {
+  const buttons = templateButtons.reduce((result: any, button: any) => {
+    const {
+      type: buttonType,
+      value,
+      title,
+      text,
+      form_id,
+      navigate_screen,
+      urlType: buttonUrlType,
+      sampleSuffix: buttonSampleSuffix,
+    }: any = button;
+
+    const urlType = buttonUrlType || 'Static';
+    const sampleSuffix = buttonSampleSuffix || '';
+
+    if (templateType?.id === CALL_TO_ACTION) {
+      const typeObj: any = {
+        phone_number: 'PHONE_NUMBER',
+        url: 'URL',
+      };
+      let obj: any = { type: typeObj[buttonType], text: title, [buttonType]: value };
+
+      if (buttonType === 'url' && urlType === 'Dynamic') {
+        obj = {
+          type: typeObj[buttonType],
+          text: title,
+          [buttonType]: `${value}{{1}}`,
+          example: [`${value}${sampleSuffix}`],
+        };
+      }
+      result.push(obj);
+    }
+
+    if (templateType?.id === QUICK_REPLY) {
+      const obj: any = { type: QUICK_REPLY, text: value };
+      result.push(obj);
+    }
+
+    if (templateType?.id === 'WHATSAPP_FORM') {
+      const obj = { type: 'FLOW', navigate_screen, text, flow_id: form_id, flow_action: 'NAVIGATE' };
+      result.push(obj);
+    }
+    return result;
+  }, []);
+
+  const templateBody = getTemplateAndButton(body);
+  const templateExample = getTemplateAndButton(getExampleFromBody(body, variables));
+
+  return {
+    hasButtons: true,
+    buttons: JSON.stringify(buttons),
+    buttonType: templateType?.id,
+    body: templateBody.message,
+    example: templateExample.message,
+  };
+};
+
+export interface BuildTemplatePayloadContext {
+  isEditing: boolean;
+  category: any;
+  variables: Array<any>;
+  isAddButtonChecked: boolean;
+  templateType: any;
+  templateButtons: Array<any>;
+  body: string;
+  footer: string;
+  tagId: any;
+}
+
+// the setPayload handed to FormLayout — turns the raw Formik values into the
+// SessionTemplateInput shape the create/update mutations expect.
+export const buildTemplatePayload = (
+  payload: any,
+  {
+    isEditing,
+    category,
+    variables,
+    isAddButtonChecked,
+    templateType,
+    templateButtons,
+    body,
+    footer,
+    tagId,
+  }: BuildTemplatePayloadContext
+) => {
+  let payloadCopy = { ...payload, isHsm: true };
+  if (isEditing) {
+    payloadCopy.shortcode = payloadCopy.newShortcode;
+  } else if (payload.languageVariant) {
+    // HSM.tsx's "translate existing HSM" flow: reuse an existing shortcode
+    // (picked from a dropdown) instead of typing a brand new one.
+    payloadCopy.category = category.label;
+    payloadCopy.shortcode = payloadCopy.existingShortcode.label;
+  } else {
+    // HSMV2.tsx (and HSM.tsx's plain create) always create a brand new
+    // template — the "translate existing HSM" flow is the only place that
+    // reuses a shortcode, and only HSM.tsx exposes that toggle.
+    payloadCopy.category = category.label;
+    payloadCopy.shortcode = payloadCopy.newShortcode;
+  }
+  payloadCopy.languageId = payload.language.id;
+  payloadCopy.example = getExampleFromBody(payloadCopy.body, variables);
+  if (isAddButtonChecked && templateType?.id) {
+    const templateButtonData = getButtonTemplatePayload({ templateButtons, templateType, body, variables });
+    Object.assign(payloadCopy, { ...templateButtonData });
+  }
+
+  if (payloadCopy.type) {
+    payloadCopy.type = payloadCopy.type.id;
+    // STICKER is a type of IMAGE
+    if (payloadCopy.type.id === 'STICKER') {
+      payloadCopy.type = 'IMAGE';
+    }
+  } else {
+    payloadCopy.type = 'TEXT';
+  }
+
+  if (payloadCopy.type === 'TEXT' || isEditing) {
+    delete payloadCopy.attachmentURL;
+  }
+  if (footer?.trim()) {
+    payloadCopy.footer = footer.trim();
+  }
+
+  if (tagId) {
+    payloadCopy.tagId = payload.tagId.id;
+  }
+
+  delete payloadCopy.isAddButtonChecked;
+  delete payloadCopy.templateButtons;
+  delete payloadCopy.language;
+  delete payloadCopy.languageVariant;
+  delete payloadCopy.variables;
+  delete payloadCopy.existingShortcode;
+  delete payloadCopy.newShortcode;
+  return payloadCopy;
+};
+
+// the next templateButtons array after adding a new button row for the given type.
+export const buildTemplateButtonsList = (
+  templateButtons: Array<any>,
+  templateType: any,
+  addFromTemplate: boolean = true
+) => {
+  let buttons: any = [];
+  const hasPhoneNumber = templateButtons.some((btn: any) => btn.type === 'phone_number');
+  const urlCount = templateButtons.filter((btn: any) => btn?.type === 'url').length;
+  let newButton = { ...buttonTypes[templateType?.id] };
+  if (templateType?.id === CALL_TO_ACTION) {
+    if (hasPhoneNumber && urlCount < 2) {
+      newButton.type = 'url';
+      newButton.value = '';
+      newButton.title = '';
+    }
+  }
+  if (templateType?.id) {
+    buttons = addFromTemplate ? [...templateButtons, newButton] : [newButton];
+  }
+  return buttons;
+};
+
+// the next templateButtons array after editing one field of one button row.
+export const buildUpdatedButtons = (templateButtons: Array<any>, value: any, row: any, index: any, eventType: any) => {
+  let obj = { ...row };
+
+  if (eventType === 'type') {
+    obj = { type: value, title: '', value: '' };
+  } else {
+    obj[eventType] = value;
+  }
+
+  return templateButtons.map((val: any, idx: number) => (idx === index ? obj : val));
+};
+
+export interface ValidationSchemaContext {
+  t: any;
+  isAddButtonChecked: boolean;
+  templateType: any;
+  // HSM.tsx has a separate "Title" field; HSMV2.tsx derives the title from
+  // shortcode + language on the backend and has no such field.
+  hasTitleField?: boolean;
+  // HSM.tsx's "translate existing HSM" toggle — reusing an existing shortcode
+  // instead of typing a new one. HSMV2.tsx always creates brand new templates.
+  hasLanguageVariant?: boolean;
+}
+
+// the Yup schema used for validating the create/edit form (skipped entirely in edit mode).
+export const buildValidationSchema = ({
+  t,
+  isAddButtonChecked,
+  templateType,
+  hasTitleField,
+  hasLanguageVariant,
+}: ValidationSchemaContext) => {
+  const validation: any = {
+    language: Yup.object().nullable().required('Language is required.'),
+    ...(hasTitleField
+      ? { label: Yup.string().required(t('Title is required.')).max(50, t('Title length is too long.')) }
+      : {}),
+    type: Yup.object()
+      .nullable()
+      .when('attachmentURL', {
+        is: (val: string) => val && val !== '',
+        then: (schema) => schema.nullable().required(t('Type is required.')),
+      }),
+    attachmentURL: Yup.string()
+      .nullable()
+      .when('type', {
+        is: (val: any) => val && val.id,
+        then: (schema) => schema.required(t('Attachment URL is required.')),
+      }),
+    footer: Yup.string().max(60, t('Footer value can be at most 60 characters')),
+    body: Yup.string().required(t('Message is required.')).max(1024, 'Maximum 1024 characters are allowed'),
+    category: Yup.object().nullable().required(t('Category is required.')),
+    variables: Yup.array().of(
+      Yup.object().shape({
+        text: Yup.string().required('Variable is required').min(1, 'Text cannot be empty'),
+      })
+    ),
+    ...(hasLanguageVariant
+      ? {
+          newShortcode: Yup.string().when('languageVariant', {
+            is: (val: any) => val === true,
+            then: (schema) => schema.nullable(),
+            otherwise: (schema) =>
+              schema
+                .required(t('Element name is required.'))
+                .matches(regexForShortcode, 'Only lowercase alphanumeric characters and underscores are allowed.'),
+          }),
+          existingShortcode: Yup.object().when('languageVariant', {
+            is: (val: any) => val === true,
+            then: (schema) => schema.nullable().required(t('Element name is required.')),
+            otherwise: (schema) => schema.nullable(),
+          }),
+        }
+      : {
+          newShortcode: Yup.string()
+            .required(t('Element name is required.'))
+            .matches(regexForShortcode, 'Only lowercase alphanumeric characters and underscores are allowed.'),
+        }),
+    templateButtons: Yup.array().of(
+      Yup.lazy(() => {
+        if (isAddButtonChecked) {
+          if (templateType?.id === 'CALL_TO_ACTION') {
+            return Yup.object().shape({
+              type: Yup.string().required('Type is required.'),
+              title: Yup.string().required('Title is required.'),
+              value: Yup.string().required('Value is required.'),
+            });
+          } else if (templateType?.id === 'QUICK_REPLY') {
+            return Yup.object().shape({
+              value: Yup.string().required('Value is required.'),
+            });
+          } else if (templateType?.id === 'WHATSAPP_FORM') {
+            return Yup.object().shape({
+              form_id: Yup.string().required('Form is required.'),
+              text: Yup.string().required('Button title is required.'),
+              navigate_screen: Yup.string().required('Screen is required.'),
+            });
+          }
+          return Yup.object().shape({});
+        }
+        return Yup.object().shape({});
+      })
+    ),
+  };
+
+  return Yup.object().shape(validation, [['type', 'attachmentURL']]);
 };
